floating point

+++ src/parser.h
@@ -4,11 +4,11 @@
 typedef struct tree_node
 {
     token_type        operator;
-    int               value;
+    float             value;
     struct tree_node *left;
     struct tree_node *right;
 } tree_node;
 
 tree_node *parse(float parent_precedence);
 tree_node *parse_leaf();
-int        evaluate(arena *arena, char *exp);
+float      evaluate(arena *arena, char *exp);
